+++ InnOut/libs/cocos2d/CCParticleExamples.m
@@ -2,7 +2,6 @@
  * cocos2d for iPhone: http://www.cocos2d-iphone.org
  *
  * Copyright (c) 2008-2010 Ricardo Quesada
- * Copyright (c) 2011 Zynga Inc.
  * 
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -40,7 +39,7 @@
 	return [self initWithTotalParticles:1500];
 }
 
--(id) initWithTotalParticles:(NSUInteger)p
+-(id) initWithTotalParticles:(int)p
 {
 	if( (self=[super initWithTotalParticles:p]) ) {
 		// duration
@@ -117,7 +116,7 @@
 	return [self initWithTotalParticles:250];
 }
 
--(id) initWithTotalParticles:(NSUInteger) p
+-(id) initWithTotalParticles:(int) p
 {
 	if( (self=[super initWithTotalParticles:p]) ) {
 
@@ -197,7 +196,7 @@
 	return [self initWithTotalParticles:350];
 }
 
--(id) initWithTotalParticles:(NSUInteger) p
+-(id) initWithTotalParticles:(int) p
 {
 	if( (self=[super initWithTotalParticles:p]) ) {
 
@@ -277,7 +276,7 @@
 	return [self initWithTotalParticles:200];
 }
 
--(id) initWithTotalParticles:(NSUInteger)p
+-(id) initWithTotalParticles:(int)p
 {
 	if( (self=[super initWithTotalParticles:p]) ) {
 
@@ -360,7 +359,7 @@
 	return [self initWithTotalParticles:250];
 }
 
--(id) initWithTotalParticles:(NSUInteger) p
+-(id) initWithTotalParticles:(int) p
 {
 	if( (self=[super initWithTotalParticles:p]) ) {
 	
@@ -443,7 +442,7 @@
 	return [self initWithTotalParticles:150];
 }
 
--(id) initWithTotalParticles:(NSUInteger) p
+-(id) initWithTotalParticles:(int) p
 {
 	if( (self=[super initWithTotalParticles:p]) ) {
 
@@ -526,7 +525,7 @@
 	return [self initWithTotalParticles:500];
 }
 
--(id) initWithTotalParticles:(NSUInteger) p
+-(id) initWithTotalParticles:(int) p
 {
 	if( (self=[super initWithTotalParticles:p]) ) {
 	
@@ -609,7 +608,7 @@
 	return [self initWithTotalParticles:700];
 }
 
--(id) initWithTotalParticles:(NSUInteger)p
+-(id) initWithTotalParticles:(int)p
 {
 	if( (self=[super initWithTotalParticles:p]) ) {
 	
@@ -691,7 +690,7 @@
 	return [self initWithTotalParticles:200];
 }
 
--(id) initWithTotalParticles:(NSUInteger) p
+-(id) initWithTotalParticles:(int) p
 {
 	if( (self=[super initWithTotalParticles:p]) ) {
 	
@@ -767,7 +766,7 @@
 	return [self initWithTotalParticles:700];
 }
 
--(id) initWithTotalParticles:(NSUInteger)p
+-(id) initWithTotalParticles:(int)p
 {
 	if( (self=[super initWithTotalParticles:p]) ) {
 	
@@ -849,7 +848,7 @@
 	return [self initWithTotalParticles:1000];
 }
 
--(id) initWithTotalParticles:(NSUInteger)p
+-(id) initWithTotalParticles:(int)p
 {
 	if( (self=[super initWithTotalParticles:p]) ) {
